added tcp master
